+++ 99.wc_object/wc.rb
@@ -1,12 +1,7 @@
+#!/usr/bin/env ruby
+
 # frozen_string_literal: true
 
-class Wc
-  attr_reader :line_count, :word_count, :byte_count, :name
+require_relative './wc_app'
 
-  def initialize(text, filename = '')
-    @line_count = text.split("\n").size
-    @word_count = text.split.size
-    @byte_count = text.bytesize
-    @name = filename
-  end
-end
+FileStat.display(ARGV)
