--- 99.wc_object/wc_command.rb
@@ -1,7 +1,0 @@
-#!/usr/bin/env ruby
-
-# frozen_string_literal: true
-
-require_relative './file_stat'
-
-FileStat.display(ARGV)
